[APM]chore(all):refactor the apm project
