+++ .gitignore
@@ -11,3 +11,4 @@
 *.obj
 *.idb
 *.pch
+*.user
