--- FFT_1/FFT_1.sln
@@ -1,20 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 10.00
-# Visual Studio 2008
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "FFT_1", "FFT_1\FFT_1.vcproj", "{ACDD27B7-FAA4-4774-B9EB-E41778CB29BE}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Win32 = Debug|Win32
-		Release|Win32 = Release|Win32
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{ACDD27B7-FAA4-4774-B9EB-E41778CB29BE}.Debug|Win32.ActiveCfg = Debug|Win32
-		{ACDD27B7-FAA4-4774-B9EB-E41778CB29BE}.Debug|Win32.Build.0 = Debug|Win32
-		{ACDD27B7-FAA4-4774-B9EB-E41778CB29BE}.Release|Win32.ActiveCfg = Release|Win32
-		{ACDD27B7-FAA4-4774-B9EB-E41778CB29BE}.Release|Win32.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
