+++ apm_process/apm_proc.c
@@ -1,0 +1,164 @@
+#include "apm_proc.h"
+
+const double Hamming_Win[FRAMELEN]={
+    0.080000,0.080089,0.080357,0.080803,0.081427,0.082229,0.083209,0.084365,
+    0.085699,0.087209,0.088894,0.090755,0.092789,0.094997,0.097378,0.099931,
+    0.102654,0.105547,0.108609,0.111837,0.115232,0.118792,0.122515,0.126400,
+    0.130446,0.134650,0.139012,0.143529,0.148200,0.153023,0.157997,0.163118,
+    0.168385,0.173797,0.179351,0.185045,0.190876,0.196843,0.202943,0.209174,
+    0.215533,0.222018,0.228626,0.235355,0.242202,0.249165,0.256241,0.263426,
+    0.270719,0.278117,0.285616,0.293214,0.300907,0.308693,0.316569,0.324532,
+    0.332578,0.340705,0.348909,0.357187,0.365536,0.373953,0.382434,0.390977,
+    0.399577,0.408231,0.416937,0.425690,0.434488,0.443327,0.452203,0.461113,
+    0.470054,0.479022,0.488014,0.497026,0.506054,0.515096,0.524147,0.533205,
+    0.542265,0.551324,0.560379,0.569426,0.578462,0.587482,0.596485,0.605465,
+    0.614420,0.623346,0.632239,0.641097,0.649916,0.658692,0.667422,0.676103,
+    0.684730,0.693302,0.701814,0.710264,0.718647,0.726961,0.735202,0.743368,
+    0.751455,0.759460,0.767380,0.775211,0.782951,0.790597,0.798146,0.805595,
+    0.812940,0.820180,0.827311,0.834331,0.841236,0.848024,0.854693,0.861240,
+    0.867663,0.873958,0.880124,0.886157,0.892057,0.897820,0.903444,0.908927,
+    0.914266,0.919461,0.924509,0.929407,0.934154,0.938749,0.943188,0.947472,
+    0.951597,0.955562,0.959367,0.963008,0.966486,0.969798,0.972943,0.975920,
+    0.978729,0.981367,0.983834,0.986128,0.988250,0.990198,0.991971,0.993568,
+    0.994990,0.996235,0.997303,0.998194,0.998907,0.999442,0.999799,0.999978,
+    0.999978,0.999799,0.999442,0.998907,0.998194,0.997303,0.996235,0.994990,
+    0.993568,0.991971,0.990198,0.988250,0.986128,0.983834,0.981367,0.978729,
+    0.975920,0.972943,0.969798,0.966486,0.963008,0.959367,0.955562,0.951597,
+    0.947472,0.943188,0.938749,0.934154,0.929407,0.924509,0.919461,0.914266,
+    0.908927,0.903444,0.897820,0.892057,0.886157,0.880124,0.873958,0.867663,
+    0.861240,0.854693,0.848024,0.841236,0.834331,0.827311,0.820180,0.812940,
+    0.805595,0.798146,0.790597,0.782951,0.775211,0.767380,0.759460,0.751455,
+    0.743368,0.735202,0.726961,0.718647,0.710264,0.701814,0.693302,0.684730,
+    0.676103,0.667422,0.658692,0.649916,0.641097,0.632239,0.623346,0.614420,
+    0.605465,0.596485,0.587482,0.578462,0.569426,0.560379,0.551324,0.542265,
+    0.533205,0.524147,0.515096,0.506054,0.497026,0.488014,0.479022,0.470054,
+    0.461113,0.452203,0.443327,0.434488,0.425690,0.416937,0.408231,0.399577,
+    0.390977,0.382434,0.373953,0.365536,0.357187,0.348909,0.340705,0.332578,
+    0.324532,0.316569,0.308694,0.300907,0.293214,0.285616,0.278117,0.270719,
+    0.263426,0.256241,0.249165,0.242202,0.235355,0.228626,0.222018,0.215533,
+    0.209174,0.202943,0.196843,0.190876,0.185045,0.179351,0.173797,0.168385,
+    0.163118,0.157997,0.153023,0.148200,0.143529,0.139012,0.134650,0.130446,
+    0.126400,0.122515,0.118792,0.115232,0.111838,0.108609,0.105547,0.102654,
+    0.099931,0.097378,0.094998,0.092789,0.090755,0.088894,0.087209,0.085699,
+    0.084365,0.083209,0.082229,0.081427,0.080803,0.080357,0.080089,0.080000};
+
+int open_pcm_data(pcm_data *pcm_data_ptr)
+{
+    int err = 0;
+
+    if(NULL == pcm_data_ptr) 
+    {
+        return -1;
+    }
+ 
+    err = fopen_s(&pcm_data_ptr->input_ptr,"origin_audio.pcm","rb");  
+
+    if(0!=err)
+    {
+        return -2;
+    }
+ 
+    err = fopen_s(&pcm_data_ptr->output_ptr,"output.txt","w+");
+
+    if(0!=err)
+    {
+        return -2;
+    }
+    err = fopen_s(&pcm_data_ptr->output_file_ptr,"output.pcm","wb");
+
+    if(0!=err)
+    {
+        return -2;
+    }
+
+    return 0;
+}
+
+
+
+
+int close_pcm_data(pcm_data *pcm_data_ptr)
+{
+    if( NULL == pcm_data_ptr 
+        || NULL == pcm_data_ptr->input_ptr
+        ||NULL == pcm_data_ptr->output_ptr
+        ||NULL == pcm_data_ptr->output_file_ptr)
+    {
+        return -1;
+    }
+    fclose(pcm_data_ptr->input_ptr);
+    fclose(pcm_data_ptr->output_ptr);
+    fclose(pcm_data_ptr->output_file_ptr);
+
+    return 0;
+}
+
+int proc_pcm_data(pcm_data *pcm_data_ptr) 
+{
+    short recover_array[160];
+    int i;
+
+    if (NULL == pcm_data_ptr)
+	{
+		return -1;
+	}
+    
+    memset(pcm_data_ptr->in_array,0,sizeof(short)*FRAMELEN); 
+    memset(pcm_data_ptr->out_array,0,sizeof(short)*FRAMELEN); 
+    memset(pcm_data_ptr->tmp_array,0,sizeof(short)*FRAMELEN); 
+    memset(pcm_data_ptr->pre_out_array,0,sizeof(short)*FRAMELEN); 
+    memset(recover_array,0,sizeof(short)*(FRAMELEN-STEP)); 
+
+     while (!feof(pcm_data_ptr->input_ptr))   
+    {
+       
+        memcpy(pcm_data_ptr->in_array,&pcm_data_ptr->in_array[STEP],sizeof(short)*(FRAMELEN-STEP));  
+        fread(&pcm_data_ptr->in_array[STEP],sizeof(short),FRAMELEN-STEP,pcm_data_ptr->input_ptr); 
+
+        for(i=0;i<FRAMELEN;i++)
+        {
+            fprintf(pcm_data_ptr->output_ptr,"%d,",pcm_data_ptr->in_array[i]); //分帧后数据
+            printf("%d", pcm_data_ptr->in_array[i]);
+
+            pcm_data_ptr->tmp_array[i]=(short)(pcm_data_ptr->in_array[i]*Hamming_Win[i]); //加窗后数据进行了short截断
+         }
+         fprintf(pcm_data_ptr->output_ptr,"\n"); 
+
+
+       complexSet(samples, pcm_data_ptr->tmp_array, FRAMELEN);
+       dft(XFreq, samples, FRAMELEN,pcm_data_ptr->out_array);
+
+
+        for(i=0;i<FRAMELEN;i++)
+        {
+
+            if(pcm_data_ptr->out_array[i]>0)
+            {
+          
+                printf("%d   %d\n",i,pcm_data_ptr->out_array[i] );
+            }
+
+        }
+
+       idft(XFreq, xTime,FRAMELEN,pcm_data_ptr->tmp_array);
+
+
+       
+        for(i=0;i<FRAMELEN-STEP;i++)
+        {
+            
+            recover_array[i] = pcm_data_ptr->pre_out_array[i+160]+pcm_data_ptr->tmp_array[i];
+            
+        }
+        for(i=0;i<FRAMELEN;i++)
+        {
+            pcm_data_ptr->pre_out_array[i] = pcm_data_ptr->tmp_array[i];
+        }
+
+        fwrite(recover_array, sizeof(short),FRAMELEN-STEP ,pcm_data_ptr->output_file_ptr);
+       
+
+     }  
+
+   return 0;    
+}
