+++ apm_process/apm_process.sln
@@ -1,0 +1,20 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 10.00
+# Visual Studio 2008
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "apm_process", "apm_process.vcproj", "{2B9B0201-AFA8-4694-A94A-FDAE5B13ED57}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Win32 = Debug|Win32
+		Release|Win32 = Release|Win32
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{2B9B0201-AFA8-4694-A94A-FDAE5B13ED57}.Debug|Win32.ActiveCfg = Debug|Win32
+		{2B9B0201-AFA8-4694-A94A-FDAE5B13ED57}.Debug|Win32.Build.0 = Debug|Win32
+		{2B9B0201-AFA8-4694-A94A-FDAE5B13ED57}.Release|Win32.ActiveCfg = Release|Win32
+		{2B9B0201-AFA8-4694-A94A-FDAE5B13ED57}.Release|Win32.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
