+++ apm_process/defines.h
@@ -1,0 +1,7 @@
+#ifndef DEFINES_H
+#define DEFINES_H
+
+#define FRAMELEN 320
+#define STEP 160
+
+#endif
