+++ apm_process/frame.c
@@ -1,0 +1,11 @@
+#include "frame.h"
+
+void frame()
+{
+    return;
+}
+
+void recover()
+{
+    return;
+}
