+++ apm_process/frame.h
@@ -1,0 +1,9 @@
+#ifndef FRAME_H
+#define FRAME_H
+
+#include "defines.h"
+
+void frame();
+void recover();
+
+#endif
